requirements.txt
Flask==0.10.1
Flask-RESTful==0.2.11
Jinja2==2.7.2
MarkupSafe==0.18
Pillow==2.3.0
Werkzeug==0.9.4
aniso8601==0.82
argparse==1.2.1
boto==2.25.0
gunicorn==18.0
ipdb==0.8
ipython==1.2.0
itsdangerous==0.23
pymongo==2.6.3
pytz==2013.9
requests==2.2.1
six==1.5.2
wsgiref==0.1.2
